# Optimize analysis backend